refactor: improved modularity, camelcased names
