+++ client/src/Components/EventCard/EventCard.scss
