+++ client/src/Pages/EventDetails/EventDetails.scss
