+++ client/src/Routes/Routes.js
@@ -5,7 +5,7 @@ import history from './History'
 import App from '../App'
 import MyEvents from '../Pages/MyEvents/MyEvents.js';
 import AddEvent from '../Pages/AddEvent/AddEvent.js';
-import EventDetails from '../Pages/EventDetails';
+import EventDetails from '../Pages/EventDetails/EventDetails';
 
 function Routes () {
