+++ client/src/assets/Navbar-elephant-logo.jpg
